requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
pydantic>=2.8.2
pymongo>=4.8.0
python-dotenv>=1.0.1
bcrypt>=4.2.0
redis>=5.0.8
httpx>=0.27.2
beautifulsoup4>=4.12.3
mcp>=1.0.0
elasticsearch>=8.15.1
langchain-core>=1.0.0
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.3.0
pypdf>=4.3.1
